Truncated icd cohort, it's currently set to a very low 10 to speed up query development in 01a-cohorts-icd-codes
Developed query-mulitplexer to aid fetching of data with multiple stay id's this reduces the query number from
10s of thousand to less than 100, still takes significant time to do the full run, circa 30mins. Could multi thread,
but I think it is bandwidth dependent, so wouldn't want to throw more than a handful of threads at it.

Made a duplicate of data preprocessing notebook to prove that the query multiplexer works, can fetch mean arterial pressure,
but this needs to be incorporated into a proper notebook. Didn't want to step on Akash's toes incase he was also working on the
data fetching notebook this weekend, most of the work was outside the notebook in query_multiplexer.py

bp data gets serialised in bp_results.out